requirements.txt
-i https://pypi.org/simple
alembic==1.0.7
awscli==1.16.111
awsebcli==3.14.11
bcrypt==3.1.6
beautifulsoup4==4.7.1
blessed==1.15.0
blinker==1.4
boto3==1.9.101
botocore==1.12.101
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
click==7.0
colorama==0.3.9
django-htmlmin==0.10.0
docker-compose==1.23.2
docker-pycreds==0.4.0
docker==3.7.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
flask-assets==0.12
flask-bcrypt==0.7.1
flask-caching==1.5.0
flask-login==0.4.1
flask-mail==0.9.1
flask-migrate==2.4.0
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
flask==1.0.2
get==2018.11.19
html5lib==1.0.1
idna==2.7
itsdangerous==1.1.0
jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
mako==1.0.7
markupsafe==1.1.1
pathspec==0.5.5
pillow==5.4.1
post==2018.11.20
psycopg2-binary==2.7.7
psycopg2==2.7.7
public==2018.11.20
pyasn1==0.4.5
pycparser==2.19
python-dateutil==2.8.0 ; python_version >= '2.7'
python-dotenv==0.10.1
python-editor==1.0.4
pyyaml==3.13
query-string==2018.11.20
request==2018.11.20
requests==2.20.1
rsa==3.4.2
s3transfer==0.2.0
semantic-version==2.5.0
six==1.11.0
soupsieve==1.8
sqlalchemy==1.2.18
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.1 ; python_version >= '3.4'
wcwidth==0.1.7
webassets==0.12.1
webencodings==0.5.1
websocket-client==0.55.0
werkzeug==0.14.1
wtforms==2.2.1
